requirements.txt
discord.py
colorama
dataset
pypubsub
stuf